+++ modules/core/quick/src/Plugin/QuickForm/QuickFormBase.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Drupal\farm_quick\Plugin\QuickForm;
 
 use Drupal\Core\Access\AccessResult;
+use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\Messenger\MessengerInterface;
 use Drupal\Core\Messenger\MessengerTrait;
@@ -30,6 +31,20 @@ class QuickFormBase extends PluginBase implements QuickFormInterface, ContainerF
   protected string $quickId;
 
   /**
+   * The entity type manager service.
+   *
+   * @var \Drupal\Core\Entity\EntityTypeManagerInterface
+   */
+  protected $entityTypeManager;
+
+  /**
+   * Current user object.
+   *
+   * @var \Drupal\Core\Session\AccountInterface
+   */
+  protected $currentUser;
+
+  /**
    * Constructs a QuickFormBase object.
    *
    * @param array $configuration
@@ -38,11 +53,17 @@ class QuickFormBase extends PluginBase implements QuickFormInterface, ContainerF
    *   The plugin_id for the plugin instance.
    * @param mixed $plugin_definition
    *   The plugin implementation definition.
+   * @param \Drupal\Core\Entity\EntityTypeManagerInterface $entity_type_manager
+   *   The entity type manager service.
+   * @param \Drupal\Core\Session\AccountInterface $current_user
+   *   Current user object.
    * @param \Drupal\Core\Messenger\MessengerInterface $messenger
    *   The messenger service.
    */
-  public function __construct(array $configuration, $plugin_id, $plugin_definition, MessengerInterface $messenger) {
+  public function __construct(array $configuration, $plugin_id, $plugin_definition, EntityTypeManagerInterface $entity_type_manager, AccountInterface $current_user, MessengerInterface $messenger) {
     parent::__construct($configuration, $plugin_id, $plugin_definition);
+    $this->entityTypeManager = $entity_type_manager;
+    $this->currentUser = $current_user;
     $this->messenger = $messenger;
   }
 
@@ -54,7 +75,9 @@ class QuickFormBase extends PluginBase implements QuickFormInterface, ContainerF
       $configuration,
       $plugin_id,
       $plugin_definition,
-      $container->get('messenger')
+      $container->get('entity_type.manager'),
+      $container->get('current_user'),
+      $container->get('messenger'),
     );
   }