+++ modules/quick/inventory/src/Plugin/QuickForm/Inventory.php
@@ -43,25 +43,11 @@ class Inventory extends QuickFormBase implements ConfigurableQuickFormInterface 
   use QuickTermTrait;
 
   /**
-   * The entity type manager service.
-   *
-   * @var \Drupal\Core\Entity\EntityTypeManagerInterface
-   */
-  protected $entityTypeManager;
-
-  /**
    * Asset inventory service.
    *
    * @var \Drupal\farm_inventory\AssetInventoryInterface
    */
   protected $assetInventory;
-
-  /**
-   * Current user object.
-   *
-   * @var \Drupal\Core\Session\AccountInterface
-   */
-  protected $currentUser;
 
   /**
    * Constructs a QuickFormBase object.
@@ -72,21 +58,18 @@ class Inventory extends QuickFormBase implements ConfigurableQuickFormInterface 
    *   The plugin_id for the plugin instance.
    * @param mixed $plugin_definition
    *   The plugin implementation definition.
-   * @param \Drupal\Core\Messenger\MessengerInterface $messenger
-   *   The messenger service.
    * @param \Drupal\Core\Entity\EntityTypeManagerInterface $entity_type_manager
    *   The entity type manager service.
-   * @param \Drupal\farm_inventory\AssetInventoryInterface $asset_inventory
-   *   Asset inventory service.
    * @param \Drupal\Core\Session\AccountInterface $current_user
    *   Current user object.
+   * @param \Drupal\Core\Messenger\MessengerInterface $messenger
+   *   The messenger service.
+   * @param \Drupal\farm_inventory\AssetInventoryInterface $asset_inventory
+   *   Asset inventory service.
    */
-  public function __construct(array $configuration, $plugin_id, $plugin_definition, MessengerInterface $messenger, EntityTypeManagerInterface $entity_type_manager, AssetInventoryInterface $asset_inventory, AccountInterface $current_user) {
-    parent::__construct($configuration, $plugin_id, $plugin_definition, $messenger);
-    $this->messenger = $messenger;
-    $this->entityTypeManager = $entity_type_manager;
+  public function __construct(array $configuration, $plugin_id, $plugin_definition, EntityTypeManagerInterface $entity_type_manager, AccountInterface $current_user, MessengerInterface $messenger, AssetInventoryInterface $asset_inventory) {
+    parent::__construct($configuration, $plugin_id, $plugin_definition, $entity_type_manager, $current_user, $messenger);
     $this->assetInventory = $asset_inventory;
-    $this->currentUser = $current_user;
   }
 
   /**
@@ -97,10 +80,10 @@ class Inventory extends QuickFormBase implements ConfigurableQuickFormInterface 
       $configuration,
       $plugin_id,
       $plugin_definition,
-      $container->get('messenger'),
       $container->get('entity_type.manager'),
-      $container->get('asset.inventory'),
       $container->get('current_user'),
+      $container->get('messenger'),
+      $container->get('asset.inventory'),
     );
   }